+++ firmware/controllers/engine_cycle/main_trigger_callback.h
@@ -15,3 +15,4 @@ void mainTriggerCallback(uint32_t trgEventIndex, efitick_t edgeTimestamp, angle_
 
 void endSimultaneousInjection(InjectionEvent *event);
 void turnInjectionPinLow(InjectionEvent *event);
+void turnInjectionPinLowStage2(InjectionEvent* event);